requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
requests==2.32.5

# Garmin integration
garminconnect==0.2.8

# LLM integration
google-generativeai==0.8.6

# Database (MySQL + SQLAlchemy)
SQLAlchemy==2.0.23
alembic==1.12.1
PyMySQL==1.1.0

# Optional: for other LLM providers
# openai==1.3.7
# anthropic==0.7.7
